# faster term aggregation fetch terms