--- a/sstable/src/dictionary.rs
+++ b/sstable/src/dictionary.rs
@@ -341,7 +341,7 @@ impl<TSSTable: SSTable> Dictionary<TSSTable> {
     /// Returns the terms for a _sorted_ list of term ordinals.
     ///
     /// Returns true if and only if all terms have been found.
-    pub fn sorted_ords_to_term_cb<F: FnMut(&[u8])>(
+    pub fn sorted_ords_to_term_cb<F: FnMut(&[u8]) -> io::Result<()>>(
         &self,
         ord: impl Iterator<Item = TermOrdinal>,
         mut cb: F,
@@ -372,7 +372,7 @@ impl<TSSTable: SSTable> Dictionary<TSSTable> {
                 bytes.extend_from_slice(current_sstable_delta_reader.suffix());
             }
             current_ordinal = ord + 1;
-            cb(&bytes);
+            cb(&bytes)?;
         }
         Ok(true)
     }
@@ -597,19 +597,28 @@ mod tests {
         // Single term
         let mut terms = Vec::new();
         assert!(dic
-            .sorted_ords_to_term_cb(100_000..100_001, |term| { terms.push(term.to_vec()) })
+            .sorted_ords_to_term_cb(100_000..100_001, |term| {
+                terms.push(term.to_vec());
+                Ok(())
+            })
             .unwrap());
         assert_eq!(terms, vec![format!("{:05X}", 100_000).into_bytes(),]);
         // Single term
         let mut terms = Vec::new();
         assert!(dic
-            .sorted_ords_to_term_cb(100_001..100_002, |term| { terms.push(term.to_vec()) })
+            .sorted_ords_to_term_cb(100_001..100_002, |term| {
+                terms.push(term.to_vec());
+                Ok(())
+            })
             .unwrap());
         assert_eq!(terms, vec![format!("{:05X}", 100_001).into_bytes(),]);
         // both terms
         let mut terms = Vec::new();
         assert!(dic
-            .sorted_ords_to_term_cb(100_000..100_002, |term| { terms.push(term.to_vec()) })
+            .sorted_ords_to_term_cb(100_000..100_002, |term| {
+                terms.push(term.to_vec());
+                Ok(())
+            })
             .unwrap());
         assert_eq!(
             terms,
@@ -621,7 +630,10 @@ mod tests {
         // Test cross block
         let mut terms = Vec::new();
         assert!(dic
-            .sorted_ords_to_term_cb(98653..=98655, |term| { terms.push(term.to_vec()) })
+            .sorted_ords_to_term_cb(98653..=98655, |term| {
+                terms.push(term.to_vec());
+                Ok(())
+            })
             .unwrap());
         assert_eq!(
             terms,